v0.1/requirements.txt
tensorflow==2.6.0
numpy==1.19.5
matplotlib==3.2.2
stockstats==0.3.2
pandas==1.0.5
tqdm==4.59.0
ta==0.7.0
yfinance==0.1.63
Pillow
scikit_learn